updated yarn.lock
so that latest "@theia/*" extensions are pulled

Problem: Following the Theia update to support Node 10.x,
I tried to test if there were impacts on this repo here.
However I noticed that older @theia/* extensions were being
pulled, from a couple of releases ago. I think it's because
we had stepped the "minor" versions vs what was in yarn.lock.
In this case, yarn will not automatically pull the new version
for these packages, as they may be incompatible.

I first attempted to run "yarn upgrade-interactive" to update
the "@theia/*" extensions but it crashed before being done. So
in the end I just removed all references to these extensions
from "yarn.lock" and it did the trick.

Fixes #24

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>