Prototype showing variables in memory view
This patch makes it so we hightlight the local variables of the current
frame in the memory view.  The idea is simple, we get the size and
address of all locals of the current frame through GDB (sizeof(foo) and
&foo).  If the bytes backing these variables are in the scope of the
memory view, we highlight them with a different color for each variable.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>
Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>
Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>